requirements.txt
opencv-contrib-python
scikit-image
numpy~=1.21.4
imutils~=0.5.4
future~=0.18.2
Pillow~=8.4.0
clang~=5.0
dlib~=19.22.1
setuptools~=59.6.0
pip~=21.3.1
tensorflow~=2.7.0
Keras~=2.4.3
scipy~=1.7.3
scikit-learn~=1.0.1
pandas~=1.1.5
h5py~=3.1.0
matplotlib~=3.5.1
Theano~=1.0.5